test_retract_pair for file table
